+++ CHANGELOG.md
@@ -8,6 +8,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Sort artifacts by known format ([#76])
+
+[#76]: https://github.com/rojo-rbx/rokit/pull/76
+
 ## `0.2.5` - August 28th, 2024
 
 ### Added